requirements.txt
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
Flask==2.1.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.0
protobuf==4.21.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.20.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
